+++ src/entrypoints/basecamp-auth/main.ts
@@ -1,5 +1,7 @@
 import { applyI18n } from "../../shared/i18n-dom";
 import { startCountdown } from "../../shared/countdown";
+import { initLocaleOverride } from "../../shared/i18n-override";
 
+await initLocaleOverride();
 applyI18n();
 startCountdown();